add ungraded planner items to syllabus
for both the teacher and the student

fixes ADMIN-1219

test plan:
- add an assignment, a graded discussion, a calendar event, an ungraded
discussion with a to-do date, and a page with a to-do date to a course.
- view the course's syllabus as a student. All of those items should
show up with the appropriate icon.
- view the course's syllabus as the teacher. same items should show.
- set the course's home page to the syllabus. the syllabus should be
viewable as the course's home page with the ungraded items.
- make the course public and view the syllabus without logging in.
should be able to see the syllabus, but the ungraded items will not
be shown. this will be fixed in a separate ticket.
- turn off the planner feature flag and view the syllabus page and the
course's home page as a teacher and a student. the syllabus should be
shown without the ungraded items.

Change-Id: Ia2309ef31bf1e508dff06e3ecd5796a117ef7e44
Reviewed-on: https://gerrit.instructure.com/157936
Tested-by: Jenkins
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Luke Kingsley <lkingsley@instructure.com>
Product-Review: Jon Willesen <jonw+gerrit@instructure.com>