Scroll Today to the top if it and the next item fit on the screen
closes ADMIN-1242

Clicking the today button forces Today to scroll to the top of the
browser, then, if the next item is not on the screen, scroll it to
top.

In fixing this, I discovered an issue where the scroll-to-top code
was working in the browser's default timezone, not the users. That's
fixed too.

Upped test coverage.

Test plan:
  - user with an item today
  - clck Today
  > expect Today to scroll to the top of the window and the item
    to get focus
  - no items today, and an item 5 days from today
  - make the browser short enough that Today and the item can't both
    be on the screen
  - click Today button
  > expect Today to scroll to the top of the page, then the next item
    to scroll to the top and be given focus
  - no items in the future, an item in the past
  - click Today button
  > expect the item to scroll to the top of the browser and get focus

Change-Id: I1e4d7d6f83873ddfafa65e14a1734af81e136627
Reviewed-on: https://gerrit.instructure.com/157876
Tested-by: Jenkins
Reviewed-by: Jon Willesen <jonw+gerrit@instructure.com>
QA-Review: Jon Willesen <jonw+gerrit@instructure.com>
Product-Review: Christi Wruck