fix vo reading of files list header
Turns out VO was confused by having a "gridcell" in the header instead
of a "columnheader". But changing it to a columnheader caused it to read
"Select All" as the header for each row's checkbox that selects the one
file. Putting the select all checkbox in various other places caused VO
to not read it at all.

fixes ADMIN-1265

test plan:
- On the files page, should be able to view the new location of the
  "Select All" checkbox with a screenreader and keyboard navigation.
- The select all checkbox should still work.
- All screenreaders are sane and read the right table header while
  traversing a file's fields.

Change-Id: I46a27d265ef13da8e97638eab580bbaf829e8d09
Reviewed-on: https://gerrit.instructure.com/159508
Tested-by: Jenkins
Reviewed-by: Mysti Sadler <mysti@instructure.com>
Reviewed-by: Carl Kibler <ckibler@instructure.com>
Product-Review: Nathan Rogowski <nathan@instructure.com>
QA-Review: Nathan Rogowski <nathan@instructure.com>