set seconds of due_at to 59
On assignment creation, the due_at has seconds of 59. So if a user
sets the due time to 11:59, the backend has it at 11:59:59. But upon
editing an assignment, the frontend had it set to 11:59:00, which
would update the assignment to 11:59:00 potentially causing some
confusion when students tried to submit. Now, the frontend when
editing an assignment will always set the seconds to 59 because
there's no way to edit that in the UI anyway.

fixes GRADE-908

Test Plan
 - Create an assignment.
 - Set a due date and time.
 - Open a rails console and view the due_at attribute of that
   assignment.
 - Verify that the seconds is :59.
 - Edit the assignment from the assignment index page.
 - Verify in the rails console that the seconds is still :59.
 - Edit the assignment from the assignment edit page.
 - Verify in the rails console that the seconds is still :59.

Change-Id: Iadfe07f41d2282c5e50af1cfa234d697c5d3654d
Reviewed-on: https://gerrit.instructure.com/159900
Tested-by: Jenkins
Reviewed-by: Jeremy Neander <jneander@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: Adrian Packel <apackel@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>