Fix focus outline on dashboard card action buttons
Though this was recently fixed in ADMIN-1218, a bug in Safari is causing
problems. This fixes it, though at the visual expense of the focus
outline not being nicely tight around the icon.

Update: in firefox, the outline on the <a> is obscured by the jewel with the
notification count. I'm doing a bit of browser sniffing to pick which
approach to take.

closes ADMIN-1327

test plan
  - load the card dashboard
  - tab through
  > expect the focus outline to behave well on the icons across the
    bottom of each card
  - switch to the recent activity dashboard
  > expect focus indication to look right as you tab around in here too

Change-Id: Icee8044e73b13c263b5bb12199637ae6d0ff257f
Reviewed-on: https://gerrit.instructure.com/160900
Tested-by: Jenkins
Reviewed-by: Dan Minkevitch <dan@instructure.com>
QA-Review: Carl Kibler <ckibler@instructure.com>
Product-Review: Dan Sasaki <dsasaki@instructure.com>