Include context_name in planner item response
Closes: CORE-1779

this is so we can not have to load all the possible courses on the
Student planner on page load just so we can join them to course_ids
From planner_item api responses to show course names for planner items.
Because right now, for people that have many courses it is more
Work to get all possible courses than it is to get the actual planner 
items.


Test plan:
* make a request to api/v1/planner_items, the returned
  response should include context_name and context_image for items
  That are linked to a context

Change-Id: I69f8560f19b2a46bc94915d4a9e13623b518b291
Reviewed-on: https://gerrit.instructure.com/162035
Reviewed-by: James Williams  <jamesw@instructure.com>
Reviewed-by: Mysti Sadler <mysti@instructure.com>
Tested-by: Jenkins
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>