don’t load section-tab info on planner dashboard
Fixes: CORE-1789

Nothing in the planner actually uses this info. So we can avoid doing
all the work to decide to show what tabs are available to this person
by not including this. It is a big win because right now, there is no
limit to the number of courses we pass through this since 
@current_user.courses_with_primary_enrollment is limitless.

There is another commit coming that will also make the dashcards, which
do need this info, faster but at least there is a maximum of 20 courses
that runs against.

This will drastically help the load that was caused by turning on the
Student planner feature flag

Test plan:
* load the student planner dashboard
* it should be a lot faster and there should not be anything
  you missed out on

Change-Id: I5cd52cef95501d4ce607bf241060f0bf502260d2
Reviewed-on: https://gerrit.instructure.com/162640
Tested-by: Jenkins
Reviewed-by: Rob Orton <rob@instructure.com>
Reviewed-by: Mysti Sadler <mysti@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>