close module tooltips
canvas' jquery tooltip inherits from the vanilla jquery tooltip and adds
a 20ms timeout to the open sequence to let the browser scroll the
element into view before jquery tries to postion it. for some reason I
can't devine, the timeout interferes with the tooltip's lifecycle and
was causing tips to get stuck open when clicking on the button. this
didn't happen in assigments, which use the same code so I suspected it
had something to do with how long the api calls took. Testing on beta,
where the server isn't colocated with the UI, I confirmed that even on
assignments, the tooltips get stuck when clicking the publish button,
which strengthens my confidence in my theory.  Increasing the timeout to
200ms seems to work, though I came at that value by simply adding a 0.
It's fast enough that the delay isn't perceptible. Hopefully it's long
enough to work in prod. Time will tell.

Also, the 'change:published' event listener that changes the tooltip
from Publish to Published was causing the button to render twice right
in a row, and that was also confusing the tooltip code, so I removed it
and it's working much better.

closes ADMIN-1332

test plan:
  - have a course with a module with a couple objects in it
  - mouse over and out of the publish/unpublish icon button
  > expect the tooltip to come and go
  - click the button
  - mouse out
  > expect the tooltip to close
  > expect tooltips on the other buttons to open/close as expected
  - click the button again, but make sure the mouse lingers over the
    button until the tooltip changes, then mouse out
  > expect the tooltip to close and other tooltips to work
  - do that again so you test both publishing and unpublishing, because
    they are a little different

  - repeat on the assignments page to be sure it works there too

Change-Id: I00ba077fe18e33b685ee00f9b110afbe1f41c0d2
Reviewed-on: https://gerrit.instructure.com/164493
Reviewed-by: Jon Willesen <jonw+gerrit@instructure.com>
QA-Review: Jon Willesen <jonw+gerrit@instructure.com>
Tested-by: Jenkins
Product-Review: Ed Schiebel <eschiebel@instructure.com>