remove harmful characters when downloading submissions
test plan:
 - as a student, upload a file to your files
 - rename it to have a slash in its name
 - submit it to an assignment
 - as a teacher, download submissions for the assignment
   in a zip fie
 - re-upload the zip file and it should associate the file
   with the prior submission

fixes ADMIN-2391

note that the ids in the filename are what Canvas uses to
reassociate the file from the zip with its submission;
the only reason we included the name was for the teacher's benefit.
using a slash in the name puts the file in a subdirectory and
breaks the link, so we replace slashes with underscores.

Change-Id: I389088c853775f52028157db7c7c003863df889f
Reviewed-on: https://gerrit.instructure.com/180808
Tested-by: Jenkins
QA-Review: Carl Kibler <ckibler@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>
Reviewed-by: Carl Kibler <ckibler@instructure.com>