fix some Individual Gradebook filters' persistence
"Treat Ungraded as 0" now persists and changes to the setting from
either Individual Gradebook or Old Gradebook are reflected in the
other, but does nothing to New Gradebook. As a result of this commit,
New Gradebook no longer respects that setting if it was turned on in
one of the other gradebooks.

"Show Concluded Enrollments" now persists and changes to the setting
from either Individual Gradebook, New Gradebook, or Old Gradebook are
reflected in the others. Future compatibility with Old Gradebook is
not guaranteed.

closes GRADE-1861

Test Plan
- Toggle on/off "Treat Ungraded as 0" in Individual Gradebook.
- Verify that the setting value is the same in Old Gradebook.
- Verify that the setting value persists between refreshes.

- Toggle on/off "Show Concluded Enrollments" in Individual Gradebook.
- Verify that the setting value is the same in New Gradebook.
- Verify that the setting value is the same in Old Gradebook.
- Verify that the setting value persists between refreshes.

Change-Id: I8420d2dcaa0bbd80ab928cfecb9fea5293001cd9
Reviewed-on: https://gerrit.instructure.com/182853
Tested-by: Jenkins
Reviewed-by: Adrian Packel <apackel@instructure.com>
Reviewed-by: Jeremy Neander <jneander@instructure.com>
QA-Review: Indira Pai <ipai@instructure.com>
Product-Review: Keith Garner <kgarner@instructure.com>