Decrease specificity of a css selector from our reset
closes: CORE-2780

This is to decrease the specificity of this css selector so we don’t
Need to add all the bloated css that we had to add in g/181185

the way I changed it to is exactly the same as the current version
of https://necolas.github.io/normalize.css/ (the thing we originally
copy/pasted this from). So I’m assuming the “WebKit but in android 4.0”
that it was referring to has been fixed. But either way, that is not
a browser we support anyway.

By committing this, we can revert (or significantly shrink) g/181185

Test plan:
* from the browser on an android phone
* check to make sure buttons appear the same as they did before

Change-Id: Iface2fc278db4e18e236d8fef26702917db75246
Reviewed-on: https://gerrit.instructure.com/185032
Tested-by: Jenkins
Reviewed-by: Brent Burgoyne <bburgoyne@instructure.com>
QA-Review: Brent Burgoyne <bburgoyne@instructure.com>
Product-Review: Brent Burgoyne <bburgoyne@instructure.com>