show correct course in scheduler notification
when a teacher and student share multiple course enrollments and
the instructor sends a message from a calendar appointment it was
showing the course context for course with the greatest ID instead
of the course from which the appointment was configured.

we now take the context code from the timeslot when a message is
sent, rather than all contexts they share.

*note* this will only impact new conversations. because existing
conversations have saved multiple contexts we do not know which
one is intended and cannot safely retroactively correct them.

Test Plan:
* create and publish two courses
* in course A, add a student and a teacher enrollment
* in course B, add the same student and teacher as A
* as the teacher, create an appointment group linked to course A
* as the student, sign up for an appointment slot for course A
* as the teacher, navigate to the calendar page
* select the appointment and message the student
* confirm that the notification context name is for course A
* confirm that the inbox message context name is for course A
* confirm that the api message context name is for course A

* repeat the above steps, but create the appointment group for
  course B and ensure the message context is for course B instead

* repeat the above steps, but add multiple students to the courses
  and the appointment group

* confirm that messages for the same course group together in the
  same conversation
* confirm that messages for different courses show up as separate
  conversations

fixes COMMS-1989

Change-Id: I4849678a25af82999a629a18def07069c51b81a9
Reviewed-on: https://gerrit.instructure.com/189709
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Matthew Lemon <mlemon@instructure.com>
Product-Review: Ryan Norton <rnorton@instructure.com>