Use es Modules for “plugin extensions”
closes CORE-2874

This makes it so we use es Modules instead of AMD for the generated
Code we make in our webpack loader so that it works if those extensions
Are AMD OR ES Modules.  As it is written before this, if any plugin
Was written as an es module it would not work unless the 
Babel add-module-exports plugin was also used.

We have to do this before I can merge my other commits that 
AMD->ES modules stuff in plugin repos

Test plan:
* plugins that add extensions to other files should still work

Change-Id: I2b6d6ee7fcabc28864b29e7f8df336316d1a3b61
Reviewed-on: https://gerrit.instructure.com/191710
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>