Decrease the delay of seeing text on first visit
Closes: CORE-2960

Google just barely released the feature to be able to control the
`font-display` property of stuff you load using it’s api.
(See: https://twitter.com/addyosmani/status/1128548064287952896)
The way it did it before is it would block any text from being rendered
until the font was loaded. This meant that a user would see a page
but with no text on it until the font was downloaded. Especially bad
for mobile and people on slow internet connections

This change makes it so the text is rendered immediately (in lato, if
it is in the browser cache, or in the Helvetica Neue fallback on a first 
visit) so there is never any delay in seeing content,
but then renders again in lato once it is done downloading.

Design/Product reviewers:
There is another setting `font-display:optional` which would make it
so on that first page load with a clear cache you’d see everything in
Helvetica, but not re-render once the lato font is downloaded. but then
on every other page load it would use lato now that is in the cache.
You can see the difference and an explanation on 
https://font-display.glitch.me/
If you’d prefer that instead we could do that but this seems like the
most logical thing to me, but I’m happy either way

Test plan:
* open your Chrome dev tools and go to the network panel
* in the “throttling” menu, pick “slow 3g”
* clear your browser cache
* reload the page, the text should appear immediately (before lato is
  done downloading) in Helvetica and then re-render once lato is loaded
* go to another page, the text should render immediately but this time
  using lato even from the start

Change-Id: I181bb1b983c5d1dea54e317cb292907bf47a835f
Reviewed-on: https://gerrit.instructure.com/193740
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>