display 'Uploaded' step in Assignments2 pizza tracker
fixes COMMS-1999

The 'Uploaded' step in the pizza tracker corresponds to the
Submission state where the user has selected files via the
FileDrop and they are visible on the assignment, but not yet
submitted via the Submit button (effectively a draft copy). When
the React state has knowledge of some files in the FileDrop that
are not yet submitted, the pizza tracker reflects the 'Uploaded'
step. When the uploaded files are removed from the FileDrop, the
pizza tracker returns to its previous state (usually 'Available').

Test Plan:
* Create an assignment as a teacher
* Add a student to the course with the assignment
* In the rails console, add a submission draft
  to that assignment:
  ```
  s = Submission.last (or whatever query gets you the submission
      for the recently published assignment)
  draft = SubmissionDraft.create!(
    submission: s,
    submission_attempt: s.attempt
  )
  attachments = Attachment.limit 3
  draft.attachments = attachments
  s.submission_drafts = [draft]
  s.save!
  ```
* Note the student view of the assignment (in Assignments2)
  displays the pizza tracker in the 'Uploaded' step
* In the rails console, delete the submission draft for the
  assignment:
  ```
  s.submission_drafts = []
  s.save!
  ```
* Note in the student view of the assignment displays the pizza
  tracker in the 'Available' step

Change-Id: I59e3e76ac6843b82d171a1355b9995368d4dabfa
Reviewed-on: https://gerrit.instructure.com/194461
Tested-by: Jenkins
Reviewed-by: Aaron Hsu <ahsu@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>
Product-Review: Steven Burnett <sburnett@instructure.com>